+++ controllers/switchport_state_handler.go
@@ -8,7 +8,7 @@ import (
 	"github.com/metal3-io/networkconfiguration-operator/api/v1alpha1"
 	"github.com/metal3-io/networkconfiguration-operator/pkg/device/switches"
 	"github.com/metal3-io/networkconfiguration-operator/pkg/machine"
-	"github.com/metal3-io/networkconfiguration-operator/pkg/util/finalizer"
+	"github.com/metal3-io/networkconfiguration-operator/pkg/utils/finalizer"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
